feat(#79): WIP workarounds for stac-fastapi issue
WIP because stac-api-validator appears to have numerous issues. Stopped making changes to fix what it reported as validation errors while waiting for a more spec-compliant validator.
It appears that stac-fastapi forces a 4-length bbox in Collection-Item and GET search requests, when the spec should support 6-length, but reluctant to commit to this given that the validator is known to not work well. It might actually be issuing invalid requests :\